+++ src/lib/buildMockGroupFieldMap.ts
@@ -9,6 +9,7 @@ import { date } from "../model/date";
 import { embed } from "../model/embed";
 import { geoPoint } from "../model/geoPoint";
 import { image } from "../model/image";
+import { integrationFields } from "../model/integrationFields";
 import { keyText } from "../model/keyText";
 import { link } from "../model/link";
 import { linkToMedia } from "../model/linkToMedia";
@@ -29,6 +30,7 @@ const mockModelFns = {
 	embed,
 	geoPoint,
 	image,
+	integrationFields,
 	keyText,
 	link,
 	linkToMedia,