+++ src/lib/generateTags.ts
@@ -4,12 +4,20 @@ import { createFaker } from "../lib/createFaker";
 
 import { MockValueConfig } from "../types";
 
-type GenerateTagsConfig = Pick<MockValueConfig, "seed">;
+type GenerateTagsConfig = {
+	min?: number;
+	max?: number;
+} & Pick<MockValueConfig, "seed">;
 
 export const generateTags = (config: GenerateTagsConfig): string[] => {
 	const faker = createFaker(config.seed);
 
-	return Array(faker.datatype.number(2))
+	return Array(
+		faker.datatype.number({
+			min: config.min ?? 0,
+			max: config.max ?? 2,
+		}),
+	)
 		.fill(undefined)
 		.map(() =>
 			changeCase.capitalCase(