+++ src/model/integrationFields.ts
@@ -1,0 +1,22 @@
+import * as prismicT from "@prismicio/types";
+import * as changeCase from "change-case";
+
+import { createFaker } from "../lib/createFaker";
+
+import { MockModelConfig } from "../types";
+
+export type MockIntegrationFieldsModelConfig = MockModelConfig;
+
+export const integrationFields = (
+	config: MockIntegrationFieldsModelConfig = {},
+): prismicT.CustomTypeModelIntegrationFieldsField => {
+	const faker = createFaker(config.seed);
+
+	return {
+		type: prismicT.CustomTypeModelFieldType.IntegrationFields,
+		config: {
+			label: changeCase.capitalCase(faker.company.bsNoun()),
+			catalog: changeCase.snakeCase(faker.lorem.words(4)),
+		},
+	};
+};