+++ src/model/sliceZone.ts
@@ -9,34 +9,53 @@ import { MockModelConfig } from "../types";
 import { sharedSliceChoice } from "./sharedSliceChoice";
 import { slice } from "./slice";
 
-type MockSliceZoneModelConfig = {
-	choicesCount?: number;
-	withSharedSlices?: boolean;
-} & MockModelConfig;
+type MockSliceZoneModelConfig = (
+	| {
+			choices?: Record<
+				string,
+				prismicT.CustomTypeModelSlice | prismicT.CustomTypeModelSharedSlice
+			>;
+			choicesCount?: never;
+			withSharedSlices?: never;
+	  }
+	| {
+			choices?: never;
+			choicesCount?: number;
+			withSharedSlices?: boolean;
+	  }
+) &
+	MockModelConfig;
 
 export const sliceZone = (
 	config: MockSliceZoneModelConfig = {},
 ): prismicT.CustomTypeModelSliceZoneField => {
 	const faker = createFaker(config.seed);
 
-	const choicesCount =
-		config.choicesCount ?? faker.datatype.number({ min: 2, max: 6 });
-
-	const choices: Record<
+	let choices: Record<
 		string,
 		prismicT.CustomTypeModelSlice | prismicT.CustomTypeModelSharedSlice
 	> = {};
-	for (let i = 0; i < choicesCount; i++) {
-		const choiceId = generateFieldId({ seed: config.seed });
 
-		choices[choiceId] = config.withSharedSlices
-			? sharedSliceChoice()
-			: slice({ seed: config.seed });
+	if ("choices" in config) {
+		choices = config.choices || {};
+	} else {
+		const choicesCount =
+			config.choicesCount ?? faker.datatype.number({ min: 2, max: 6 });
+
+		for (let i = 0; i < choicesCount; i++) {
+			const choiceId = generateFieldId({ seed: config.seed });
+
+			choices[choiceId] = config.withSharedSlices
+				? sharedSliceChoice()
+				: slice({ seed: config.seed });
+		}
 	}
 
 	const labels: Record<string, prismicT.CustomTypeModelSliceLabel[]> = {};
-	if (!config.withSharedSlices) {
-		for (const choiceId in choices) {
+	for (const choiceId in choices) {
+		const choice = choices[choiceId];
+
+		if (choice.type === prismicT.CustomTypeModelSliceType.Slice) {
 			const labelsCount = faker.datatype.number({ min: 0, max: 3 });
 
 			labels[choiceId] = Array(labelsCount)