+++ src/types.ts
@@ -27,6 +27,8 @@ export type SetRequired<BaseType, Keys extends keyof BaseType> = Simplify<
 		Required<Pick<BaseType, Keys>>
 >;
 
+export type Seed = string | number | number[];
+
 export interface MockImageData {
 	url: string;
 	width: number;
@@ -42,8 +44,12 @@ export interface MockEmbedData {
 	thumbnail_width: number | null;
 }
 
+export type MockRestApiConfig = {
+	seed?: Seed;
+};
+
 export type MockModelConfig = {
-	seed?: number;
+	seed?: Seed;
 };
 
 // TODO: Add to @prismicio/types
@@ -55,7 +61,7 @@ export type PrismicModel =
 	| prismicT.SharedSliceModelVariation;
 
 export type MockValueConfig<Model extends PrismicModel = PrismicModel> = {
-	seed?: number;
+	seed?: Seed;
 	model?: Model;
 };
 
@@ -101,7 +107,7 @@ type CustomTypeModelStructuredTextField =
 export type MockRichTextValueConfig<
 	Model extends CustomTypeModelStructuredTextField = CustomTypeModelStructuredTextField,
 > = {
-	seed?: number;
+	seed?: Seed;
 	model?: Model;
 };
 
@@ -168,8 +174,8 @@ type CustomTypeModelFieldForGroupValue<
 	? prismicT.EmbedField
 	: T extends prismicT.CustomTypeModelGeoPointField
 	? prismicT.GeoPointField
-	: T extends prismicT.CustomTypeModelIntegrationField
-	? prismicT.IntegrationField
+	: T extends prismicT.CustomTypeModelIntegrationFieldsField
+	? prismicT.IntegrationFields
 	: never;
 
 type CustomTypeModelGroupFieldValue<