+++ test/model-sharedSlice.test.ts
@@ -32,6 +32,7 @@ test("can be configured with specific repeat and non-repeat field configuration"
 				embed: { count: 0 },
 				geoPoint: { count: 0 },
 				image: { count: 0 },
+				integrationFields: { count: 0 },
 				keyText: { count: 0 },
 				link: { count: 0 },
 				linkToMedia: { count: 0 },
@@ -51,6 +52,7 @@ test("can be configured with specific repeat and non-repeat field configuration"
 				embed: { count: 0 },
 				geoPoint: { count: 0 },
 				image: { count: 0 },
+				integrationFields: { count: 0 },
 				keyText: { count: 0 },
 				link: { count: 0 },
 				linkToMedia: { count: 0 },
@@ -73,4 +75,15 @@ test("can be configured with specific repeat and non-repeat field configuration"
 
 	t.is(primaryFieldIds.length, 1);
 	t.is(variation.primary[primaryFieldIds[0]].type, "Boolean");
+});
+
+test("can be configured to use a given set of variations", (t) => {
+	const variations = [
+		model.sharedSliceVariation(),
+		model.sharedSliceVariation(),
+	];
+
+	const actual = model.sharedSlice({ variations });
+
+	t.deepEqual(actual.variations, variations);
 });