+++ test/value-sharedSlice.test.ts
@@ -26,6 +26,7 @@ test("supports custom model", (t) => {
 				embed: { count: 0 },
 				geoPoint: { count: 0 },
 				image: { count: 0 },
+				integrationFields: { count: 0 },
 				keyText: { count: 0 },
 				link: { count: 0 },
 				linkToMedia: { count: 0 },
@@ -45,6 +46,7 @@ test("supports custom model", (t) => {
 				embed: { count: 0 },
 				geoPoint: { count: 0 },
 				image: { count: 0 },
+				integrationFields: { count: 0 },
 				keyText: { count: 0 },
 				link: { count: 0 },
 				linkToMedia: { count: 0 },
@@ -80,18 +82,52 @@ test("returns no items if model does not include items model", (t) => {
 });
 
 test("can be customized with a pattern to determine the number of items", (t) => {
-	const actualNone = value.sharedSlice({ pattern: "none" });
+	const customModel = model.sharedSlice({
+		itemsFieldConfig: {
+			configs: {
+				boolean: { count: 1 },
+				color: { count: 0 },
+				contentRelationship: { count: 0 },
+				date: { count: 0 },
+				embed: { count: 0 },
+				geoPoint: { count: 0 },
+				image: { count: 0 },
+				keyText: { count: 0 },
+				link: { count: 0 },
+				linkToMedia: { count: 0 },
+				number: { count: 0 },
+				richText: { count: 0 },
+				select: { count: 0 },
+				timestamp: { count: 0 },
+				title: { count: 0 },
+			},
+		},
+	});
+
+	const actualNone = value.sharedSlice({
+		model: customModel,
+		pattern: "none",
+	});
 	t.true(actualNone.items.length < 1);
 
-	const actualShort = value.sharedSlice({ pattern: "short" });
+	const actualShort = value.sharedSlice({
+		model: customModel,
+		pattern: "short",
+	});
 	t.true(actualShort.items.length >= 1);
 	t.true(actualShort.items.length <= 3);
 
-	const actualMedium = value.sharedSlice({ pattern: "medium" });
+	const actualMedium = value.sharedSlice({
+		model: customModel,
+		pattern: "medium",
+	});
 	t.true(actualMedium.items.length >= 3);
 	t.true(actualMedium.items.length <= 6);
 
-	const actualLong = value.sharedSlice({ pattern: "long" });
+	const actualLong = value.sharedSlice({
+		model: customModel,
+		pattern: "long",
+	});
 	t.true(actualLong.items.length >= 6);
 	t.true(actualLong.items.length <= 12);
 });