+++ test/value-slice.test.ts
@@ -22,6 +22,7 @@ test("supports custom model", (t) => {
 				embed: { count: 0 },
 				geoPoint: { count: 0 },
 				image: { count: 0 },
+				integrationFields: { count: 0 },
 				keyText: { count: 0 },
 				link: { count: 0 },
 				linkToMedia: { count: 0 },
@@ -41,6 +42,7 @@ test("supports custom model", (t) => {
 				embed: { count: 0 },
 				geoPoint: { count: 0 },
 				image: { count: 0 },
+				integrationFields: { count: 0 },
 				keyText: { count: 0 },
 				link: { count: 0 },
 				linkToMedia: { count: 0 },
@@ -76,18 +78,53 @@ test("returns no items if model does not include repeat model", (t) => {
 });
 
 test("can be customized with a pattern to determine the number of items", (t) => {
-	const actualNone = value.slice({ pattern: "none" });
+	const customModel = model.slice({
+		repeatFieldConfig: {
+			configs: {
+				boolean: { count: 1 },
+				color: { count: 0 },
+				contentRelationship: { count: 0 },
+				date: { count: 0 },
+				embed: { count: 0 },
+				geoPoint: { count: 0 },
+				image: { count: 0 },
+				integrationFields: { count: 0 },
+				keyText: { count: 0 },
+				link: { count: 0 },
+				linkToMedia: { count: 0 },
+				number: { count: 0 },
+				richText: { count: 0 },
+				select: { count: 0 },
+				timestamp: { count: 0 },
+				title: { count: 0 },
+			},
+		},
+	});
+
+	const actualNone = value.slice({
+		model: customModel,
+		pattern: "none",
+	});
 	t.true(actualNone.items.length < 1);
 
-	const actualShort = value.slice({ pattern: "short" });
+	const actualShort = value.slice({
+		model: customModel,
+		pattern: "short",
+	});
 	t.true(actualShort.items.length >= 1);
 	t.true(actualShort.items.length <= 3);
 
-	const actualMedium = value.slice({ pattern: "medium" });
+	const actualMedium = value.slice({
+		model: customModel,
+		pattern: "medium",
+	});
 	t.true(actualMedium.items.length >= 3);
 	t.true(actualMedium.items.length <= 6);
 
-	const actualLong = value.slice({ pattern: "long" });
+	const actualLong = value.slice({
+		model: customModel,
+		pattern: "long",
+	});
 	t.true(actualLong.items.length >= 6);
 	t.true(actualLong.items.length <= 12);
 });